// interfaces/artwork.ts
type Artwork = {
  Title: string
  Artist: string[]
  ConstituentID: number[]
  ArtistBio: string[]
  Nationality: string[]
  BeginDate: number[]
  EndDate: number[]
  Gender: Array<string | null>
  Date: string
  Medium: string
  Dimensions: string
  CreditLine: string
  AccessionNumber: string
  Classification: string
  Department: string
  DateAcquired: string
  Cataloged: string
  ObjectID: number
  URL: string
  ThumbnailURL: string
  'Height (cm)': number
  'Width (cm)': number
}

export default Artwork